refactor: more charts/ into pkg/fe/template/chart
This also moves the templating logic out of pkg/fe/compiler and into pkg/fe/template
