+++ pkg/fe/template/.gitignore
@@ -1,0 +1,1 @@
+*.tar.gz
