+++ pkg/fe/template/chart/templates/image-pull-secret.yaml
