+++ pkg/fe/template/chart/templates/namespace.yaml
