+++ pkg/fe/template/chart/templates/scc.yaml
