+++ pkg/lunchpail/version.go
@@ -5,7 +5,7 @@ import (
 	"strings"
 )
 
-//go:generate /bin/sh -c "[ -d ../../charts ] && grep appVersion ../../charts/Chart.yaml | tr -s ' ' | cut -d' ' -f2 > version.txt || exit 0"
+//go:generate /bin/sh -c "[ -d ../fe/template/chart ] && grep appVersion ../fe/template/chart/Chart.yaml | tr -s ' ' | cut -d' ' -f2 > version.txt || exit 0"
 //go:embed version.txt
 var version string
